+++ Dimension/Dimension_Impl/DerivedDimensions/MagneticFluxDimension_Impl.h
@@ -1,9 +1,9 @@
 #ifndef STATIC_DIMENSION_MAGNETICFLUX_IMPL_H
 #define STATIC_DIMENSION_MAGNETICFLUX_IMPL_H
 
-#include "../../LengthDimension.h"
-#include "../../MassDimension.h"
 #include "../../TimeDimension.h"
+#include "../../MassDimension.h"
+#include "../../LengthDimension.h"
 #include "../../ChargeDimension.h"
 
 namespace Dimension
@@ -73,6 +73,32 @@ namespace Dimension
 
    template<typename... Ts>
    class MagneticFlux;
+
+   /// @brief Represents a default MagneticFlux.
+   /// @details This MagneticFlux is templated on the primary units of the relevant dimensions.
+   ///   While this is a specific type, its intended use is to treat an object or parameter as an abstract
+   ///   "MagneticFlux" type, without regard for the underlying units.
+   template<>
+   class MagneticFlux<> : public BaseDimension<std::tuple<PrimaryMass, PrimaryLength, PrimaryLength>, std::tuple<PrimaryTime, PrimaryCharge>>
+   {
+   public:
+      using Base = BaseDimension<std::tuple<PrimaryMass, PrimaryLength, PrimaryLength>, std::tuple<PrimaryTime, PrimaryCharge>>;
+      using Base::Base;
+
+      /// @brief Constructs a MagneticFlux object with a value.
+      /// @param val The value of the MagneticFlux.
+      explicit constexpr MagneticFlux(PrecisionType val) : Base(val) {}
+
+      /// @brief Constructs a MagneticFlux object from another MagneticFlux object.
+      /// @tparam OtherMagneticFlux The other MagneticFlux type.
+      /// @param base The base MagneticFlux object.
+      template<typename OtherMagneticFlux>
+      requires IsMagneticFluxType<OtherMagneticFlux>
+      // Implicit conversion between dimensions of the same unit is core to Dimensional
+      // cppcheck-suppress noExplicitConstructor
+      constexpr MagneticFlux(const OtherMagneticFlux& base)
+         : Base(base.template GetVal<std::tuple<PrimaryMass, PrimaryLength, PrimaryLength>, std::tuple<PrimaryTime, PrimaryCharge>>()) {}
+   };
 
    /// @brief Represents a MagneticFlux.
    /// @details Defines operations and data storage for MagneticFlux dimensions.